Bitstream and Driver prototype code
Bitstream constructs a timeline of data and events from packets, with
the events in the place we expect them to happen.

Driver takes that and pushes the events back according to the timings
of the DMA+PWM approach, and handles the necessary repeating/looping
that has to take place as a result.

This version of Bitstream pushes a single data for many words, and
Driver slices it back up again.