Remove Cmailbox dependency
We can use ioctl() directly from Swift since this form uses the *ptr
form that a definition is provided for; we just need the ioctl number
constant out of the header, which we can make by hand for now.

Not having the C module in the mix makes Xcode and LLDB much much
happier, and that makes me much happier when debugging test failures.