Remove bit duration assumptions from Bitstream
Having unit tests means I can do something awesome; make Bitstream
truly agnostic as to the transmission details. It's now initialized
with a `bitDuration` parameter, and calculates the resulting lengths
of logical bits and the RailCom cutout from that.

This makes the RailCom code a little more fun since it has to
dynamically deal with where the markers go, and even deal with a
postamble, but that's easy to test now!