Wrap mailbox memory in an UncachedMemory class
This is something we don't want to accidentally "leak" and have to clean
up at the end of the program, and there's a whole annoying bunch of baggage
in addition to the pointer we have to carry around - the mailbox reference,
size, and handle to it.

It therefore makes sense to wrap these in a class with a deinit{}, and since
we're doing it, have busAddress as a member too.  Thus we now have

  do {
      let mem = try UncachedMemory.allocate(minimumSize: 16, using: raspberryPi)
      let array = mem.pointer.bindMemory(to: Int.self, capacity: 2)

  } // automatically cleaned up