Move /dev/mem and /dev/vcio opening later
This allows us to initialize a RaspberryPi() and have it just be the
stash of address-details for the hardware until we need to actually
map memory (which throws anyway, so is the right time to open that
device node).

For testing, we add an internal init() that just takes fixed constants,
that means we have an object we can actually use for things.

Move UncachedMemory to be a bit more of an internal implementation
detail, and have just a generic mapMemory() function instead of a
specific mapPeripheral(). These are the two that need the
memory file descriptor, and mailbox - but also things like the
pageSize, uncachedAliasAddress, etc. constants.