Start on a new Driver code with DriverTests
Still experimenting with exactly how this will go, but I basically like
the idea of Bitstream->CB+Data and then initialize the Uncached Memory
from that.

Testing can focus on parseBitstream() nicely, though I'll almost certainly
want a bunch of test helpers for the details of this. The new RaspberryPi
approach seems to work.

Needed a CBSD module for arc4random_uniform() on Linux, but was easy enough
to include and make Linux-only. Keeps things compatible with Xcode.