Parse Bitstream into DMA Control Blocks
This new code is a lot more flexible (and tested!) than the code in
OldDriver, in particular it's capable of multiple loop unrolls where
required and supports a non-repeating and repeating section within
the Bitstream.

Still TODO: the Driver code for queuing a Bitstream, waiting for the
previous bitstream to complete at least one iteartion, changing from an
existing one into the new one, and detecting when the new one has begun
so we can free the old one.