Rename ParsedBitstream to QueueBitstream and add tests for commit()
Since we're going to use the same structure before and after committing,
it makes more sense to refer to this as a queued bitstream; that means
commit() doesn't need to return the memory region, just store it in the
structure.

Add tests for commit() now we've decided for sure what it's going to do.