Move startup/shutdown/queue code to new Driver
Now seems like a good point to throw out OldDriver; we can move the
startup and shutdown code pretty much wholesale into Driver, since these
are hardware nitty-gritty there isn't a huge rush to write tests here and
will probably change over time anyway.

The queue() method needs a complete and total rewrite, and tests, but it
makes sense to at least drop in a version now and deal with that by
replacing the code in-situ.