Use a DispatchQueue to sequence post-queue checks.
After queueing a bitstream we need to wait for it to be transmitted
before removing the old bitstream (which cleans up its memory), and
we need to wait for the bitstream to repeat before calling the new
completion handler.

Sequence these by using a dispatch queue for all modifications to the
bitstream queue array; the initial add is synchronous, and then
schedules a repeated check on isTransmitting; once true it removes the
old queue item, and schedules a check on isRepeating; once that's true
we call the completion handler.

Each block captures the specific QueuedBitstream item.