Move the end control block to the real end
The end control block is used to signal completion, so should go at the
end of the bitstream data, not at the point we happen to finish unrolling.
In addition, there only needs to be one of them, so don't repeat even if
we unroll multiple copies of the loop.