Add Bitstream.breakpoint
A breakpoint is a point within a bitstream where it would be safe, after
the first full output, to break out of the bitstream and transfer to
another. One is automatically placed at the end of the bitstream, others
may be placed as makes sense.

QueuedBitstream handles Breakpoints by saving not just the controlBlocks
offset that would have its nextControlBlockAddress changed, but also the
range and delayedEvents that existed at that point.

parsedBitstream can be called with a Breakpoint, after the initial parsing,
which will append a new start control block for it, and then act as if
the loop were being unrolled from that point - in many cases this will
mean we can just jump right into the original bitstream, some will require
some unrolling.

(In theory we don't need the initial parse, and can just stack transfer ones
on top of one another - will experiment with that later).

This approach is way cleaner than all of the stitching/joining ideas I
had, since it just relies on the code we already have (cleaned up quite
a bit)-we just generate custom control blocks to do the transfer.