Remove bitstream from the QueuedBitstream initializer
Instead simply allow parseBitstream() to be called as many times as
we like; and in the transfer case, only for the breakpoints since the
"coming in fresh" scenario will never get used.

Right now we still assume the same bitstream each time, but if there's
a way to identify them, we might even be able to remove that restriction.

At the same time, remove the copied and stored properties for the bitstream
such as duration and completionHandler, though this temporarily breaks that
handler it's okay because we're not using or testing that yet.