Clean up QueuedBitstream's transfer functions
The combination of transfer(from:to) and transfer(to:at) seems to be
about as clean as we can get it, since we have to co-operate between
the two values and pass data between them with a commit() inbetween.

As a bonus, hide the whole business of handling repeating vs. non-repeating
inside this method, it turns out the best way to do it is change the
end control blocks always, and only the earlier ones after if we're
already repeating.

Replicate the same behavior for the stop() method.

These methods are simple enough to test, so add the tests.