Multithreading fixes
The assert in Driver.queue isn't accurate in the cases where the
whenRepeating blocks run out-of-order, as with the remove(at:) above,
it's enough that the count is correct.

When waiting for repeating in the tests, also push an empty sync block
since the dispatch group is left inside the block, and wait() might
complete before the block releases its captures on memory we're about
to test.

Wrap uses of markRepeating/Transmitting and DMA active changes in a
sync {} block too, to stop the thread from acting on the changes while
the test is still making them (and e.g. shortening the array and breaking
the indexes).