Start work on the RailCom Detector
The basic approach is an interrupt to watch the logic changes from an
opto-coupler, and a timer to track the deltas between them.

Currently battling with how much to go in the interrupt handler vs.
how much in the main loop. Repeatedly disabling/enabling interrupts in
order to check a flag doesn't do any more favours than an oversized ISR.