Pull signal analysis out of ISR
Pretty much a total rework of the detector code, moving everything
aside from edge delta calculation out of the ISR and into the main
loop.

Rather than a ring-buffer (which overflowed when the UART was too
slow to keep up), just check for edges at the end of the loop. If
they occur faster than we can process, we either have a bad signal
or a problem.