Do a tiny bit of code minimization
analyzing the signal is time-critical, so add an extra switch case
to avoid an if check, use a shifting bitmask to avoid repeated math
on the bit number, and roll the check byte code into the final case
to avoid duplicated ifs.

This code is pretty long, but it's also pretty unwound without much
duplication of effort.