Remove the edge paranoia
An interrupt being ready at the end of the loop doesn't mean that we
missed anything; and we can miss things simply by virtue of the
interrupt being masked while it's running and while we fetch from it.

So this really doesn't buy us anything rather than an annoying beep
whenever the RailCom cutout happens on every packet.