Make Packer a protocol, and BitPacker conform to it
Packable thus ends up being a super-protocol-oriented protocol of any
type that can add itself to a Packer.

Packer is any object that Packables can be added to.

The magic coems by providing that conformance through the adding fixed
width integers with a length that we already had, and we can even add
default conformances to things so stdlib types are Packable.

Why do this? ... because now we can implement a DCCPacker that outputs
physical bits into words, by accepting logical values as input.