Add a much-simplified version of Bitstream
This new version is a Packer! That means that we can just do this:

  let timing = BitstreamTiming(pulseWidth: 14.5)
  var bitstream = Bitstream(timing: timing)

  let packet = SpeedAndDuration(...)
  bitstream.add(packet)

All of the event stuff from the old Bitstream is removed because in
practice trailing an extra wire to the booster for the RailCom cutout
wasn't helpful, especially since it has an AVR and can just generate it
itself.

Also all of the knowledge about what constitutes a packet, such as
start bits, inter-byte bits, etc. has moved to Packet now where it
belongs.

Likewise we don't want things like preamble anymore, while that isn't
implemented anywhere yet, a really simple implementation could look
like this:

  struct Preamble : Packable {
    var timing: BitstreamTiming

    func add<T : Packer>(into packer: inout T) {
      for _ in 0..<timing.preambleLength {
        packer.add(0b1, length: 1)
      }
    }
  }

  // then just:
  bitstream.add(Preamble(timing: timing))