Add preamble to BitstreamTiming, simplify RailCom
The 14 one bit preamble is a useful constant to keep in BitstreamTiming
and is certainly where it belongs now.

Also relevant is now much we need to extend that to include the RailCom
cutout, so change the weird different lengths to a length (in ms) and then
a count of one bits (which can always be multiplied).