PulsePacker - change the type from Int to UInt32
The Swift `Int` type is actually mapped to the C `long` type, which means that it's 64-bit on all 64-bit platforms, including LP64 where `sizeof (int)` is typically 4.

For now the best way of dealing with this is to use an explicit 32-bit sized type, matching the size of the Raspberry Pi's FIFO register, but use a typealias so this can be changed later, or even replaced with a generic.