Move old code out of the way completely
Experimenting with 64-bit Debian on the RPi 3B+ has revealed that while
C `int` is 4 bytes as I expected, Swift `Int` is actually 8-bytes.

This means that all of the existing code needs porting from using `Int` to
`UInt32` or similar.

May as well use that opportunity to clean it up and copy it back one piece
at a time.