Cleaned up GPIO support
This is derived from OldRaspberryPi/GPIO.swift, but super cleaned up and
completed.

Now GPIO is a class that encompasses all GPIO pins, and can be iterated and
subscripted to access individual pins:

  gpio[17].function = .output
  gpio[17].value = true

It's a class because it wraps an underlying pointer resource, which will need
to be cleaned up on deinitialization, and none of the usual value type
behavior applies either.

The intermediate GPIOPin class that handles the subscripting is also a class
because it just wraps two values (class, number), none of the value type
behaviors make sense, but also because by being a class it allows subscripts
to work through the reference rather than needing to be mutating which makes
little sense.

Edge and level detection is also supported now:

  gpio[17].edgeDetect = .both
  while !gpio[17].isEventDetected {}
  gpio[17].isEventDetected = false

Note that the isEventDetected actually works inverted compared to the
underlying register, because this way looks more programmatically correct.

Also the pull-up/down registers are supported, but only in so far as
setting the hardware. The reason for this is because we don't want to
introduce usleep() or delays into this code, and don't want to have
complicated callbacks or threads, etc.

Callers can implement using their own timing:

  gpio.pullUpDownEnable = .pullUp
  usleep(150) // 1MHz
  gpio[17].pullUpDownClock = true
  usleep(150)
  gpio.pullUpDownEnable = .disabled
  gpio[17].pullUpDownClock = false

And a final note, this isn't backed by a hardware map yet; this simply
backs to a pointer which is enough for testing.