Implement hardware support for GPIO
I've been struggling for ages on how to remove the RaspberryPi class/struct
that really made everything untidy to initialize and use. Dave Abrahams
always says "start with a protocol", and that's what works perfectly here!

MappedRegisters encapsulates any hardware-backed behavior we need to map
from /dev/mem via the registers area, and provides the implementation for
the common parts.

RaspberryPi thus just becomes a set of constants, which we wrap in an enum
to provide a namespace, but nothing more.

RaspberryPiError gets renamed to OSError since it's just a wrapper around
errno.