Make MappedRegisters non-static
While static works for GPIO, it won't work when we come to things like
Clock where the offset can vary based on the initializer, so we want it
to be an instance member.

Things were only static anyway because I was working around the Swift
initializer rules. The alternate way was to make registers an
ImplicitlyUnwrappedOptional instead, which is what I've done here.