Rework Clock in the style of GPIO (and now PWM)
The top-level class has a bit of a goofy name: Clocks, I might still fix
that later.

Now we do:
  let clock = try Clocks()
  clock[.pwm].isEnabled = true

This has a pleasing symmetry with GPIO and PWM now, and also means that
Clocks just does a single giant map of all of the registers - thus removing
the need for all the yucky offset mucking around as we tried to just map
one register pair.