Replace enum in GPIO.Registers with UInt32
Swift enums do not have the width of their RawValue, the compiler instead
synthesizes .rawValue based on the underlying storage tag. So this actually
has a size of 1 not 4

While on a little endian RPi this is fine, to be safe, just use the enum
to obtain the rawValue like we do everywhere else.