Copy over the Mailbox and UncachedMemory code
Do this slightly differently, since we only ever intend to use Mailbox
for this one thing, narrow its scope to be rather less sprawling.

Then have UncachedMemory to be an RAII approach that completely wraps
and hides it:

  let uncached = try UncachedMemory(minimumSize: 1024)
  // use uncached.pointer

  uncached.deallocate()

Since we have to `try` anyway, just have that do the open of the Mailbox
and clean itself up each time. there's not a huge benefit to keeping it
open - same with the /dev/mem we open.