DMATests - fix pointer usage
Caught by Xcode 11.4; in Swift the lifetime of a pointer is only valid for scope, so capturing the address of member and expecting it to outlive that call is undefined behavior, e.g. &foo is only valid for the call of that initializer. Switch to using allocate/initialize of structures and pass them in directly.