Work around Bus Error
For some reason, despite everything being aligned and strided correctly for UInt32, we get a Bus Error unless data is a multiple of Int's wider stride. I suspect a Swift bug, but can't replicate it with a minimal test case yet, and can't rule out an error in this code somewhere.

So work around it for now and plan to avoid working around it when this code is replaced.