Start some new AVR code for Booster
The older code used way too many instructions for timing, the mistake
was using a generic running clock and trying to do math, instead of
specializing the clock for the exact timing we need.

Start afresh (but will copy and paste where appropriate).

Current status is the code records the length of the high and low
periods, and detects loss of signal.